+++ Steamodded/MystBlinds/MystBlinds.lua
@@ -6,7 +6,7 @@
 --- MOD_DESCRIPTION: A pack of Blinds
 --- DISPLAY_NAME: MystBlinds
 --- BADGE_COLOUR: EDA9D3
---- VERSION: 1.2.1
+--- VERSION: 1.2.2
 
 local blind_list = {
     "market",
@@ -24,7 +24,14 @@ local blind_list = {
 }
 
 local mod_path = SMODS.current_mod.path
-SMODS.Atlas({ key = "MystBlinds", atlas_table = "ANIMATION_ATLAS", path = "MystBlinds.png", px = 34, py = 34, frames = 21 })
+SMODS.Atlas({ 
+    key = "MystBlinds", 
+    atlas_table = "ANIMATION_ATLAS", 
+    path = "MystBlinds.png", 
+    px = 34, 
+    py = 34, 
+    frames = 21 
+})
 
 -- basically taken from 5CEBalatro lol
 for k, v in ipairs(blind_list) do
@@ -36,6 +43,7 @@ for k, v in ipairs(blind_list) do
     else
         blind.key = v
         blind.atlas = "MystBlinds"
+        -- very hardcoded but w/e
         blind.pos = { x = 0, y = k - 1 }
 
         local blind_obj = SMODS.Blind(blind)
@@ -48,6 +56,24 @@ for k, v in ipairs(blind_list) do
     end
 end
 
+-- Add mod's icon
+SMODS.Atlas({
+    key = "modicon",
+    path = "mystblinds_icon.png",
+    px = 34,
+    py = 34
+}):register()
+
 sendInfoMessage("Loaded MystBlinds~")
 
 ---- Other functions ----
+
+-- Blacklist The Stone on Cryptid's enhanced decks
+local apply_to_runref = Back.apply_to_run
+function Back.apply_to_run(self)
+    apply_to_runref(self)
+
+    if self.effect.config.cry_force_enhancement then
+        G.GAME.bosses_used["bl_myst_stone"] = 1e308
+    end
+end